Secured Api's &  Detached User From Services & Attached to Controllers

+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/controller/CredentialController.java
@@ -2,8 +2,10 @@ package com.udacity.jwdnd.course1.cloudstorage.controller;
 
 
 import com.udacity.jwdnd.course1.cloudstorage.model.CredentialForm;
+import com.udacity.jwdnd.course1.cloudstorage.model.User;
 import com.udacity.jwdnd.course1.cloudstorage.services.CredentialService;
 import com.udacity.jwdnd.course1.cloudstorage.services.EncryptionService;
+import com.udacity.jwdnd.course1.cloudstorage.services.UserService;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -17,20 +19,32 @@ import javax.servlet.http.HttpServletRequest;
 @Controller
 public class CredentialController {
     private final CredentialService credentialService;
+    private final UserService userService;
 
-    public CredentialController(CredentialService credentialService, EncryptionService encryptionService) {
+    public CredentialController(CredentialService credentialService, EncryptionService encryptionService, UserService userService) {
         this.credentialService = credentialService;
+        this.userService = userService;
     }
 
     @PostMapping("/upsert-credential")
-    public ModelAndView handleUpsertCredential(@ModelAttribute("newCredential") CredentialForm credentialForm, Model model, Authentication authentication, HttpServletRequest request) {
+    public ModelAndView handleUpsertCredential(@RequestParam("userid") String userid,
+                                               @ModelAttribute("newCredential") CredentialForm credentialForm,
+                                               Model model,
+                                               Authentication authentication,
+                                               HttpServletRequest request) {
+
         request.setAttribute("tab", "credentials");
         try {
-            credentialService.upsertCredential(credentialForm, authentication.getName());
-            model.addAttribute("Credentials", credentialService.getUserCredentials(authentication.getName()));
-            request.setAttribute("errorMessage", "null");
+            User user = userService.getUser(authentication.getName());
+            if (Integer.parseInt(userid) == user.getUserid()) {
+                credentialService.upsertCredential(credentialForm, user.getUserid());
+                model.addAttribute("Credentials", credentialService.getUserCredentials(user.getUserid()));
+                request.setAttribute("errorMessage", "null");
 
-            return new ModelAndView("/result");
+                return new ModelAndView("/result");
+            } else {
+                throw new SecurityException("You are not permitted to perform this action!");
+            }
 
         } catch (Exception e) {
             request.setAttribute("errorMessage", e.getMessage());
@@ -39,18 +53,28 @@ public class CredentialController {
     }
 
     @GetMapping("/delete-credential")
-    public ModelAndView handleDeleteCredential(@RequestParam("credentialid") String credentialid, Model model, Authentication authentication, HttpServletRequest request) {
-        request.setAttribute("tab", "credentials");
-        try {
-            credentialService.deleteCredential(credentialid);
-            model.addAttribute("Credentials", credentialService.getUserCredentials(authentication.getName()));
-            request.setAttribute("errorMessage", "null");
-            return new ModelAndView("result");
+    public ModelAndView handleDeleteCredential(@RequestParam("userid") String userid,
+                                               @RequestParam("credentialid") String credentialid,
+                                               Model model,
+                                               Authentication authentication,
+                                               HttpServletRequest request) {
 
-        } catch (Exception e) {
-            request.setAttribute("errorMessage", e.getMessage());
-            return new ModelAndView("result");
-        }
+        request.setAttribute("tab", "credentials");
+            try {
+                User user = userService.getUser(authentication.getName());
+                if (Integer.parseInt(userid) == user.getUserid()) {
+                    credentialService.deleteCredential(credentialid);
+                    model.addAttribute("Credentials", credentialService.getUserCredentials(user.getUserid()));
+                    request.setAttribute("errorMessage", "null");
+                    return new ModelAndView("result");
+                } else {
+                    throw new SecurityException("You are not permitted to perform this action!");
+                }
+
+            } catch (Exception e) {
+                request.setAttribute("errorMessage", e.getMessage());
+                return new ModelAndView("result");
+            }
     }
 
 }
