+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/controller/FileController.java
@@ -2,7 +2,9 @@ package com.udacity.jwdnd.course1.cloudstorage.controller;
 
 
 import com.udacity.jwdnd.course1.cloudstorage.model.File;
+import com.udacity.jwdnd.course1.cloudstorage.model.User;
 import com.udacity.jwdnd.course1.cloudstorage.services.FileService;
+import com.udacity.jwdnd.course1.cloudstorage.services.UserService;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpHeaders;
@@ -23,13 +25,16 @@ import java.io.IOException;
 @Controller()
 public class FileController {
     private final FileService fileService;
+    private final UserService userService;
 
-    public FileController(FileService fileService) {
+    public FileController(FileService fileService, UserService userService) {
         this.fileService = fileService;
+        this.userService = userService;
     }
 
     @PostMapping("/file-upload")
     public ModelAndView handleFileUpload(
+            @RequestParam("userid") String userid,
             @RequestParam("fileUpload") MultipartFile fileUpload,
             Model model,
             Authentication authentication,
@@ -41,12 +46,15 @@ public class FileController {
             request.setAttribute("errorMessage", "No File Found");
             return new ModelAndView("/result");
         } else {
-
+            User user = userService.getUser(authentication.getName());
             try {
-                int fileId = fileService.saveFile(fileUpload, authentication.getName());
-                model.addAttribute("Files", fileService.getFiles(authentication.getName()));
-                request.setAttribute("errorMessage", "null");
-                return new ModelAndView("/result");
+                if (Integer.parseInt(userid) == user.getUserid()) {
+                    int fileId = fileService.upsertFile(fileUpload, user.getUserid());
+                    model.addAttribute("Files", fileService.getFiles(user.getUserid()));
+                    request.setAttribute("errorMessage", "null");
+                    return new ModelAndView("/result");
+                } else { throw new SecurityException("You are not permitted to perform this action!");}
+
             } catch (Exception e) {
                 request.setAttribute("errorMessage", e.getMessage());
                 return new ModelAndView("/result");
@@ -56,23 +64,35 @@ public class FileController {
     }
 
     @GetMapping("/file-download")
-    public ResponseEntity<Resource> handleFileDownload(@RequestParam(value = "fileId") String fileId) throws IOException {
-        File file = fileService.getFile(fileId);
-        return ResponseEntity.ok()
-                .contentType(MediaType.parseMediaType(file.getContenttype()))
-                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + file.getFilename() + "\"")
-                .body(new ByteArrayResource(file.getFiledata().readAllBytes()));
-
+    public ResponseEntity<Resource> handleFileDownload(@RequestParam("userid") String userid,@RequestParam(value = "fileId") String fileId, Authentication authentication) {
+        try {
+            User user = userService.getUser(authentication.getName());
+            if (Integer.parseInt(userid) == user.getUserid()) {
+                File file = fileService.getFile(fileId);
+                return ResponseEntity.ok()
+                        .contentType(MediaType.parseMediaType(file.getContenttype()))
+                        .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + file.getFilename() + "\"")
+                        .body(new ByteArrayResource(file.getFiledata().readAllBytes()));
+            } else { throw new SecurityException("You are not permitted to perform this action!");}
+        } catch (Exception ignored) {
+            return null;
+        }
     }
 
     @GetMapping("/file-delete")
-    public ModelAndView handleFileDelete(@RequestParam(value="fileId") String fileId, Authentication authentication, Model model, HttpServletRequest request) {
+    public ModelAndView handleFileDelete(@RequestParam("userid") String userid,@RequestParam(value="fileId") String fileId, Authentication authentication, Model model, HttpServletRequest request) {
         request.setAttribute("tab", "files");
         try {
-            var deleteId = fileService.deleteFile(fileId);
-            model.addAttribute("Files", fileService.getFiles(authentication.getName()));
-            request.setAttribute("errorMessage", "null");
-            return new ModelAndView("/result");
+            User user = userService.getUser(authentication.getName());
+            if (Integer.parseInt(userid) == user.getUserid()) {
+                var deleteId = fileService.deleteFile(fileId);
+                model.addAttribute("Files", fileService.getFiles(user.getUserid()));
+                request.setAttribute("errorMessage", "null");
+                return new ModelAndView("/result");
+            } else {
+                throw new SecurityException("You are not permitted to perform this action!");
+            }
+
         } catch (Exception e) {
             request.setAttribute("errorMessage", e.getMessage());
             return new ModelAndView("/result");
